requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.27.0
Django==4.2.11
django-filter==24.1
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
fsspec==2024.3.0
gTTS==2.5.1
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.0.2
inflection==0.5.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.18.1
SpeechRecognition==3.10.1
pydub==0.25.1
pocketsphinx==5.0.3
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0